# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechNest Shop API
  version: 1.0.0
  description: >
    Versioned OpenAPI contract for the current TechNest Shop backend. This file is
    the authoritative machine-readable API contract; README keeps the operational
    snapshot and docs/API_CONTRACT.md keeps a concise human summary.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a customer account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getAuthMe
      summary: Get current auth status or current user profile
      security: []
      responses:
        '200':
          description: Anonymous auth status or authenticated user profile
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthStatusResponse'
                  - $ref: '#/components/schemas/UserProfileResponse'
    put:
      tags: [Auth]
      operationId: updateAuthMe
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/categories:
    get:
      tags: [Catalog]
      operationId: listCategories
      summary: List categories
      security: []
      responses:
        '200':
          description: Categories sorted by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags: [Catalog]
      operationId: createCategory
      summary: Create category
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Catalog]
      operationId: getCategory
      summary: Get category by id
      security: []
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Catalog]
      operationId: updateCategory
      summary: Update category
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Catalog]
      operationId: deleteCategory
      summary: Delete category
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Category deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Request validation failed or payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, status, error]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]
    AuthStatusResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required: [authenticated]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
          maxLength: 100
        fullName:
          type: string
          maxLength: 120
      required: [email, password, fullName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
          maxLength: 100
      required: [email, password]
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 120
        username:
          type: string
          pattern: '^\\s*$|^[A-Za-z0-9._-]{3,50}$'
          maxLength: 50
        phone:
          type: string
          pattern: '^\\s*$|^[0-9+()\\- ]{8,20}$'
        addressText:
          type: string
          maxLength: 255
        avatarUrl:
          type: string
          maxLength: 100000
        newPassword:
          type: string
          minLength: 6
          maxLength: 100
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        username:
          type: string
        fullName:
          type: string
        phone:
          type: string
        addressText:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
        message:
          type: string
      required: [id, email, username, role]
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
      required: [id, name]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
      required: [name]
...